Изменение : Исправлена ошибка невозможности/ошиюок открытия роликов с Youtube, размером больше 4 Гб.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@5021 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/apps/mplayerc/MainFrm.h
@@ -1197,8 +1197,8 @@ private:
 		TH_CLOSE
 	};
 	volatile TH_STATE	m_fYoutubeThreadWork;
-	volatile DWORD		m_YoutubeCurrent;
-	volatile DWORD		m_YoutubeTotal;
+	volatile QWORD		m_YoutubeCurrent;
+	volatile QWORD		m_YoutubeTotal;
 	CString				m_YoutubeFile;
 	CWinThread*			m_YoutubeThread;
